Change Algorithm for MaxTotalJobs and MaxWaitingJobs
This fixes https://github.com/ic-hep/GridPPDIRAC/issues/13
This issue is that we want to change the algorithm for calculating the MaxTotalJobs and MaxWaitingJobs config variables so that MaxTotalJobs is set to GlueCEInfoTotalCPUs unless this is zero, in which case it is set to GlueSubClusterLogicalCPUs. MaxWaitingJobs we now want set to 2* MaxTotalJobs.