Allowing the banning of resources
This fixes https://github.com/ic-hep/DIRAC/issues/48 and makes use of the BannedCEs/BannedSEs config parameters which were previously ignored. Now banning a CE will also remove it from a config if it already exists.

Note that since we dont have a cleanup step for SEs yet, this is not true for banned SEs.
	modified:   ConfigurationSystem/Agent/AutoBdii2CSAgent.py
	modified:   ConfigurationSystem/private/AddResourceAPI.py