Remove duplicate Travis CI builds in PR status.
Travis triggers builds when a commit is pushed to any branch by default and also
attempts to build a PR. This causes two builds on a pull request status. These builds are slightly
different but probably only the PR one is what you want. The solution is to specify only the master branch
is built on push but that pull requests are still built.