Auto vac cs agent (#61)
* Add AutoVac2CSAgent.py

* Fix to remove old CEs from AddResoruces.py

Remember to clear the old_ces set for each site.

	modified:   ConfigurationSystem/private/AddResourceAPI.py

* Fix bugs found in production testing.

This PR fixes the following:

1) Relative import
2) config defauls for poll time
3) adding country code onto the site name
4) fix for the ConfigurationSystem helper to make append/append_unique work

The issue here is that as the changes are only pushed to the config with a commit, append and append_unique had to get the current config values from the DIRAC config before appending to either list or set. This means that two such successive calls will end up with the same old values as the new values from the first call aren't flushed to config yet. Solution is to hold onto such changes and flush them all at the end.

	modified:   Agent/AutoVac2CSAgent.py
	modified:   ConfigTemplate.cfg
	modified:   private/AutoResourceTools/ConfigurationSystem.py

* Add Agent.execute return.